+++ genhub/crg.py
@@ -149,11 +149,10 @@ class CrgDB(genhub.genomedb.GenomeDB):
 
 
 def test_scaffolds():
-    """CRG scaffolds download"""
-    config = genhub.test_registry.genome('Dqcr')
+    """CRG: scaffolds download"""
+    dqua_db = genhub.test_registry.genome('Dqcr')
     testurl = 'http://wasp.crg.eu/DQUA.v01.fa.gz'
     testpath = './Dqcr/DQUA.v01.fa.gz'
-    dqua_db = CrgDB('Dqcr', config)
     assert dqua_db.gdnaurl == testurl, \
         'scaffold URL mismatch\n%s\n%s' % (dqua_db.gdnaurl, testurl)
     assert dqua_db.gdnapath == testpath, \
@@ -162,11 +161,10 @@ def test_scaffolds():
 
 
 def test_annot():
-    """CRG annotation download"""
-    config = genhub.test_registry.genome('Dqcr')
+    """CRG: annotation download"""
+    dqua_db = genhub.test_registry.genome('Dqcr', workdir='CRG')
     testurl = 'http://wasp.crg.eu/DQUA.v01.gff3'
     testpath = 'CRG/Dqcr/DQUA.v01.gff3.gz'
-    dqua_db = CrgDB('Dqcr', config, workdir='CRG')
     assert dqua_db.gff3url == testurl, \
         'annotation URL mismatch\n%s\n%s' % (dqua_db.gff3url, testurl)
     assert dqua_db.gff3path == testpath, \
@@ -174,11 +172,10 @@ def test_annot():
 
 
 def test_proteins():
-    """CRG protein download"""
-    config = genhub.test_registry.genome('Dqcr')
+    """CRG: protein download"""
+    dqua_db = genhub.test_registry.genome('Dqcr', workdir='/opt/db/genhub')
     testurl = 'http://wasp.crg.eu/DQUA.v01.pep.fa.gz'
     testpath = '/opt/db/genhub/Dqcr/DQUA.v01.pep.fa.gz'
-    dqua_db = CrgDB('Dqcr', config, workdir='/opt/db/genhub')
     assert dqua_db.proturl == testurl, \
         'protein URL mismatch\n%s\n%s' % (dqua_db.proturl, testurl)
     assert dqua_db.protpath == testpath, \
@@ -187,8 +184,7 @@ def test_proteins():
 
 def test_protids():
     """CRG: extract protein IDs from GFF3"""
-    conf = genhub.test_registry.genome('Dqcr')
-    db = CrgDB('Dqcr', conf)
+    db = genhub.test_registry.genome('Dqcr')
     protids = ['DQUA011a006022P1', 'DQUA011a006023P1', 'DQUA011a006024P1']
     infile = 'testdata/gff3/dqua-275.gff3'
     testids = list()
@@ -201,8 +197,7 @@ def test_protids():
 
 def test_protmap():
     """CRG: extract protein-->iLocus mapping from GFF3"""
-    conf = genhub.test_registry.genome('Dqcr')
-    db = CrgDB('Dqcr', conf)
+    db = genhub.test_registry.genome('Dqcr')
     mapping = {'DQUA011a006022P1': 'DquaILC-14465',
                'DQUA011a006023P1': 'DquaILC-14466',
                'DQUA011a006024P1': 'DquaILC-14467'}
@@ -217,6 +212,6 @@ def test_protmap():
 
 def test_format():
     """GenomeDB task drivers"""
-    conf = genhub.test_registry_supp.genome('Pccr')
-    pcan_db = CrgDB('Pccr', conf, workdir='testdata/demo-workdir')
-    pcan_db.format(logstream=None)
+    db = genhub.test_registry_supp.genome('Pccr',
+                                          workdir='testdata/demo-workdir')
+    db.prep(logstream=None)