+++ genhub/iloci.py
@@ -160,10 +160,9 @@ def prepare(db, delta=500, ilcformat='%sILC-', logstream=sys.stderr):  # pragma:
 
 
 def test_intervals():
-    """Parse iLocus intervals"""
-    config = genhub.test_registry.genome('Bdis')
-    d = genhub.refseq.RefSeqDB('Bdis', config, workdir='testdata/demo-workdir')
-    intervals(d, logstream=None)
+    """iLoci: compute iLocus intervals"""
+    db = genhub.test_registry.genome('Bdis', workdir='testdata/demo-workdir')
+    intervals(db, logstream=None)
 
     outfile = 'testdata/demo-workdir/Bdis/Bdis.iloci.gff3'
     testfile = 'testdata/gff3/bdis-iloci.gff3'
@@ -175,10 +174,9 @@ def test_intervals():
 
 
 def test_simple():
-    """Determine simple iLoci"""
-    config = genhub.test_registry.genome('Bdis')
-    d = genhub.refseq.RefSeqDB('Bdis', config, workdir='testdata/demo-workdir')
-    simple(d, logstream=None)
+    """iLoci: determine simple iLoci"""
+    db = genhub.test_registry.genome('Bdis', workdir='testdata/demo-workdir')
+    simple(db, logstream=None)
 
     outfile = 'testdata/demo-workdir/Bdis/Bdis.simple-iloci.txt'
     testfile = 'testdata/misc/bdis-simple.txt'
@@ -186,10 +184,9 @@ def test_simple():
 
 
 def test_reps():
-    """Select representative gene models for each iLocus"""
-    config = genhub.test_registry.genome('Bdis')
-    d = genhub.refseq.RefSeqDB('Bdis', config, workdir='testdata/demo-workdir')
-    representatives(d, logstream=None)
+    """iLoci: select representative gene models for each iLocus"""
+    db = genhub.test_registry.genome('Bdis', workdir='testdata/demo-workdir')
+    representatives(db, logstream=None)
 
     outfile = 'testdata/demo-workdir/Bdis/Bdis.ilocus.mrnas.gff3'
     testfile = 'testdata/gff3/bdis-reps.gff3'
@@ -197,10 +194,9 @@ def test_reps():
 
 
 def test_sequences():
-    """Extract iLocus sequences"""
-    config = genhub.test_registry.genome('Bdis')
-    d = genhub.refseq.RefSeqDB('Bdis', config, workdir='testdata/demo-workdir')
-    sequences(d, logstream=None)
+    """iLoci: extract iLocus sequences"""
+    db = genhub.test_registry.genome('Bdis', workdir='testdata/demo-workdir')
+    sequences(db, logstream=None)
 
     outfile = 'testdata/demo-workdir/Bdis/Bdis.iloci.fa'
     testfile = 'testdata/fasta/bdis-iloci.fa'
@@ -212,10 +208,9 @@ def test_sequences():
 
 
 def test_ancillary():
-    """Process ancillary data for iLoci"""
-    config = genhub.test_registry.genome('Bdis')
-    d = genhub.refseq.RefSeqDB('Bdis', config, workdir='testdata/demo-workdir')
-    ancillary(d, logstream=None)
+    """iLoci: process ancillary data for iLoci"""
+    db = genhub.test_registry.genome('Bdis', workdir='testdata/demo-workdir')
+    ancillary(db, logstream=None)
 
     outfile = 'testdata/demo-workdir/Bdis/Bdis.ilens.tsv'
     testfile = 'testdata/misc/bdis-ilens.tsv'