+++ scripts/genhub-pilocus-summary.py
@@ -88,11 +88,9 @@ def main(args):
     if args.cfgdir:
         for cfgdirpath in args.cfgdir.split(','):
             registry.update(cfgdirpath)
-    conf = registry.genomes(args.species)
 
     for species in args.species:
-        config = conf[species]
-        db = genhub.genomedb.GenomeDB(species, config, workdir=args.workdir)
+        db = registry.genome(species, workdir=args.workdir)
         iloci = pandas.read_table(db.ilocustable)
         premrnas = pandas.read_table(db.premrnatable)
         row = get_row(iloci, premrnas, args.outfmt)